% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_attacking_zones.R
\name{compute_attacking_zones}
\alias{compute_attacking_zones}
\title{Compute distribution of attacks on zones}
\usage{
compute_attacking_zones(df, n_zones = 4, pitchLength = 120, pitchWidth = 80)
}
\arguments{
\item{df}{event data set}

\item{n_zones}{Number of partitions of final third}

\item{pitchLength}{Length of pitch in data set}

\item{pitchWidth}{Width of pitch in data set}
}
\value{
A dataframe containing zone borders and attacks per zone for both teams
}
\description{
Compute distribution of attacks on zones
}
\details{
An attack is every passing or carrying event that starts in the first
two thirds and ends up in any zone of the final third. Here, the end location
of the event determines to which zone the attack is assigned.
}
\examples{
data(eventData)
compute_attacking_zones(df = eventData)
}
