% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xT.R
\name{compute_trans_mat}
\alias{compute_trans_mat}
\title{Compute transition matrices}
\usage{
compute_trans_mat(df, bins_x = 16, bins_y = 12)
}
\arguments{
\item{df}{Event data set in SPADL format with bin memberships for coordinates,
as returned by \link{prepare_data}.}

\item{bins_x}{Number of bins in \code{x} direction}

\item{bins_y}{}
}
\value{
A tibble with \code{bins_x} times \code{bins_y} rows, containing a tibble with
the transition probabilities for each combination of x bin and y bin in a column
named \code{trans_prob}.
}
\description{
Compute transition matrices
}
\examples{
data(spadlData)
prep_dat <- prepare_data(spadlData)
compute_trans_mat(prep_dat$prep_data)
}
