% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xT.R
\name{compute_xT}
\alias{compute_xT}
\title{Compute xT values}
\usage{
compute_xT(df, bins_x = 16, bins_y = 12, itermax = 100, eps = 1e-05)
}
\arguments{
\item{df}{Event data set in SPADL format with bin memberships for coordinates,
as returned by \link{prepare_data}.}

\item{bins_x}{Number of bins in \code{x} direction}

\item{bins_y}{}

\item{itermax}{The maximal number of iterations to carry out}

\item{eps}{The precision to calculate the xT value of a cell; iterations are
carried out until precision or maximal number of iteration is reached}
}
\value{
A list with elements
\itemize{
\item \code{xT_final}: a tibble containing ball moving, shooting and scoring probability as
well as the expected threat for each cell.
\item \code{heatmaps} A list with a heatmap for each iteration.
}
}
\description{
Compute xT values
}
\examples{
data(spadlData)
prep_dat <- prepare_data(spadlData)
xT <- compute_xT(prep_dat$prep_data, itermax = 50, eps = 1e-03)
xT$xT_final

## plot heatmap of xT:
xT_tib <- prep_dat$bin_centers \%>\% left_join(xT$xT_final)

xT_tib \%>\% ggplot(aes( x = x_center, y = y_center, fill = xT)) +
  geom_tile(width = 105/16, height = 68/12, colour = "gray")+
  scale_fill_gradient(low = "white", high = "darkred")+
  draw_pitch(dimensions = c(105, 68), fill = NA)+
  geom_text(aes(label = round(xT, 2)))+ theme(axis.title = element_blank())

# visualise the 5 iterations:

plotlist <- purrr::map2(xT$heatmaps, 1:length(xT$heatmaps), ~ {
 prep_dat$bin_centers \%>\% left_join(.x, by = c("bin_start_x", "bin_start_y")) \%>\%
 ggplot(aes( x = x_center, y = y_center, fill = xT)) +
  geom_tile(width = 105/16, height = 68/12, colour = "gray")+
  scale_fill_gradient(low = "white", high = "darkred")+
  draw_pitch(dimensions = c(105, 68), fill = NA)+
  geom_text(aes(label = round(xT, 2)))+ theme(axis.title = element_blank())+
  ggtitle(paste("xT after", .y, "iterations"))
})
ggpubr::ggarrange(plotlist = plotlist, common.legend = TRUE)

##### with less cells: 12 in x direction, 8 in y direction

prep_dat <- prepare_data(spadlData, bins_x = 12, bins_y = 8)
xT <- compute_xT(prep_dat$prep_data, itermax = 100, eps = 1e-03, bins_x = 12, bins_y = 8)
xT$xT_final

## plot heatmap of xT:
xT_tib <- prep_dat$bin_centers \%>\% left_join(xT$xT_final)

xT_tib \%>\% ggplot(aes( x = x_center, y = y_center, fill = xT)) +
  geom_tile(width = 105/12, height = 68/8, colour = "gray")+
  scale_fill_gradient(low = "white", high = "darkred")+
  draw_pitch(dimensions = c(105, 68), fill = NA)+
  geom_text(aes(label = round(xT, 3)))+ theme(axis.title = element_blank())
}
