% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pitch.R
\name{draw_pitch}
\alias{draw_pitch}
\title{annotate a pitch to a ggplot object}
\usage{
draw_pitch(
  dimensions = c(120, 80),
  palette = "gw",
  color = NULL,
  fill = NULL,
  coord_flip = FALSE,
  reverse_scale = NULL
)
}
\arguments{
\item{dimensions}{pitch dimensions}

\item{palette}{colour palette, one of "gw", "bw", "classic", "smurf", "dark",
"wc"}

\item{color}{line colour}

\item{fill}{fill colour}

\item{coord_flip}{whether to flip coordinates}

\item{reverse_scale}{whether in the ggplot object to which the pitch is annotated
any scale was reversed. NULL if this is not the case, anything else if it is the case}
}
\value{
a list with ggplot commands
}
\description{
copied from package scoutr
}
\examples{
ggplot() + draw_pitch()
}
