% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventData.R
\docType{data}
\name{eventData}
\alias{eventData}
\title{Example Event Data set}
\format{
\subsection{\code{eventData}}{

A data frame with 3,338 rows and 156 columns.
The original event data was loaded with the \code{StatsBombR} package,
and the \code{allclean} function has been called.
}
}
\source{
\url{https://github.com/statsbomb/open-data}
}
\usage{
eventData
}
\description{
Event Data from CL final 2012/2013, publicly available from StatsBomb
}
\keyword{datasets}
