% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_danger_passes.R
\name{get_danger_passes}
\alias{get_danger_passes}
\title{Get tibble with dangerous passes}
\usage{
get_danger_passes(df, shot_window = 15, team, pitch_dim = c(120, 80))
}
\arguments{
\item{df}{Event data set}

\item{shot_window}{the time window before a shot within which passes are considered
as dangerous}

\item{team}{name of the team for which to compute dangerous passes}
}
\value{
a tibble containing information on location of dangerous passes
}
\description{
Get tibble with dangerous passes
}
\examples{
data(eventData)
get_danger_passes(df = eventData, shot_window = 20, team = "Borussia Dortmund")
}
