% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_heatmap.R
\name{pass_heatmap}
\alias{pass_heatmap}
\title{Passing heatmap}
\usage{
pass_heatmap(
  df,
  team,
  LeftToRight = TRUE,
  pitch_dim = c(120, 80),
  xBins = 8,
  yBins = 5
)
}
\arguments{
\item{df}{Event data set}

\item{team}{Name of the team for which to make a passing heatmap}

\item{LeftToRight}{Logical; whether you want the team to play from left to right
in the visualisation}

\item{pitch_dim}{Dimension of the pitch}

\item{xBins}{Number of bins in x direction}

\item{yBins}{Number of bins in y direction}
}
\value{
A ggplot object: pass heatmap for successful and failed passes along with
average passing direction
}
\description{
Passing heatmap
}
\examples{
data(eventData)
pass_heatmap(eventData, team = "Borussia Dortmund")
pass_heatmap(eventData, team = "Bayern Munich", LeftToRight = FALSE, xBins = 6, yBins =4)
}
