% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_attacking_zones.R
\name{plot_attack_zones}
\alias{plot_attack_zones}
\title{Visualise attacking zones}
\usage{
plot_attack_zones(
  dfAttack,
  team,
  pitchLength = 120,
  pitchWidth = 80,
  leftToRight = TRUE,
  scale_lim = c(0, 60)
)
}
\arguments{
\item{dfAttack}{data frame with information on zones and attacks as returned by
\code{compute_attacking_zones}}

\item{team}{Team's name for which to visualise attacking zones}

\item{pitchLength}{Length of pitch in data set}

\item{pitchWidth}{Width of pitch in data set}

\item{leftToRight}{Logical; whether the team is attacking from left to right}

\item{scale_lim}{limits on colour scale for arrow colour}
}
\value{
A ggplot object
}
\description{
Visualise attacking zones
}
\examples{
data(eventData)
dfatt <- compute_attacking_zones(eventData)
plot_attack_zones(dfatt, team = "Borussia Dortmund")
plot_attack_zones(dfatt, team = "Bayern Munich", leftToRight = FALSE)
}
