% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_danger_passes.R
\name{plot_dangerpass_heatmap}
\alias{plot_dangerpass_heatmap}
\title{Heatmap of dangerous passes}
\usage{
plot_dangerpass_heatmap(
  dfDangerPass,
  leftToRight = TRUE,
  shot_window,
  xBins = 6,
  yBins = 4,
  pitch_dim = c(120, 80),
  ...
)
}
\arguments{
\item{dfDangerPass}{A data frame containing information on dangerous passes
as returned by \code{get_danger_passes}}

\item{leftToRight}{Logical; whether you want the team to play from left to right
in the plots}

\item{shot_window}{the time window before a shot within which passes are considered
as dangerous}

\item{xBins}{the number of bins in x direction}

\item{yBins}{the number of bins in y direction}

\item{pitch_dim}{dimension of the pitch}

\item{...}{additional arguments passed to \code{ggplot}, such as \code{legend.key.width},
\code{fill_limits} (limit of fill colour scale)}
}
\value{
A list of ggplots, one with pass positions and a heatmap
}
\description{
Heatmap of dangerous passes
}
\examples{

data(eventData)
df_dp <- get_danger_passes(df = eventData, shot_window = 20, team = "Borussia Dortmund")
plot_dangerpass_heatmap(df_dp, shot_window = 20, fill_limits = c(1, 10))
df_dp <- get_danger_passes(df = eventData, shot_window = 20, team = "Bayern Munich")
plot_dangerpass_heatmap(df_dp, shot_window = 20, fill_limits = c(1, 15), leftToRight = FALSE)
}
