% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xg_shotmaps.R
\name{plot_shotmaps}
\alias{plot_shotmaps}
\title{Visualise shots with xG/outcome}
\usage{
plot_shotmaps(df, hometeam = NULL, pitch_dim = c(120, 80))
}
\arguments{
\item{df}{Event data set}

\item{hometeam}{Name of hometeam (plays from left to right)}

\item{pitch_dim}{dimension of the pitch}
}
\value{
a list of ggplot objects
}
\description{
Visualise shots with xG/outcome
}
\examples{
data(eventData)
shotmaps <- plot_shotmaps(eventData)
shotmaps$shot_types
shotmaps$shot_xg
}
