% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poss_lost_heatmap.R
\name{poss_loss_heatmap}
\alias{poss_loss_heatmap}
\title{Heatmap of possession losses}
\usage{
poss_loss_heatmap(
  df,
  team,
  xBins = 8,
  yBins = 5,
  pitch_dim = c(120, 80),
  LeftToRight = TRUE
)
}
\arguments{
\item{df}{Event data set}

\item{team}{Name of the team for which to make a passing heatmap}

\item{xBins}{Number of bins in x direction}

\item{yBins}{Number of bins in y direction}

\item{pitch_dim}{Dimension of the pitch}

\item{LeftToRight}{Logical; whether you want the team to play from left to right
in the visualisation}
}
\value{
A ggplot object
}
\description{
Heatmap of possession losses
}
\examples{
data(eventData)
poss_loss_heatmap(eventData, "Borussia Dortmund")
poss_loss_heatmap(eventData, "Bayern Munich", LeftToRight = FALSE, xBins = 6, yBins = 4)
}
