% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot_formations.R
\name{prep_formation_data}
\alias{prep_formation_data}
\title{Prepare a data frame to visualise tactical formation}
\usage{
prep_formation_data(
  df,
  team.name,
  player_var_name = "player.name",
  pos_form_var_name = "position.name",
  pos_id_var_name = "position.id",
  jerseynum_var_name = "jersey_number"
)
}
\arguments{
\item{df}{A data frame as contained in statsbomb data in \code{tactics.lineup},
containing information on player names, position name, position id according to
statsbomb's definition, jersey numbers}

\item{team.name}{The team's name}

\item{player_var_name}{Variable/column name of player names}

\item{pos_form_var_name}{Variable/column name of position name}

\item{pos_id_var_name}{Variable/column name of position id}

\item{jerseynum_var_name}{Variable/column name of jersey number}
}
\value{
A data frame, columns are renamed and locations of positions are added
}
\description{
Prepare a data frame to visualise tactical formation
}
\examples{
 # see help of `plot_formation`
}
