% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xT.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for computing xT}
\usage{
prepare_data(
  df,
  pitch_dim = list(from = c(105, 68), to = c(105, 68)),
  reverse_y = TRUE,
  bins_x = 16,
  bins_y = 12
)
}
\arguments{
\item{df}{An event data set in SPADL format}

\item{pitch_dim}{A list with items
*\code{from} original pitch dimension (numeric vector with \code{x} and \code{y} coordinate,
in that order)
*\code{to} pitch dimension to convert to (numeric vector with \code{x} and \code{y} coordinate,
in that order)}

\item{reverse_y}{logical; whether to reverse the \code{y}-values (if upper left corner of
pitch is set to 0 instead of e.g. 68)}

\item{bins_x}{Number of bins in \code{x}-direction}

\item{bins_y}{Number of bins in \code{y}-direction}
}
\value{
A list with items
\itemize{
\item \code{prep_data} The prepped event data with columns \code{start_x}, \code{start_y}, \code{end_x},
\code{end_y}, \code{actiontype}, \code{result}, \code{bin_start_x}, \code{bin_start_y}, \code{bin_end_x}, \code{bin_end_y},
with coordinates transformed (if required) and respective bins values for coordinates
\item \code{bin_centers} All combinations of bins and their centers. Useful for plotting.
}
}
\description{
Prepare data for computing xT
}
\examples{
data(spadlData)
prep_dat <- prepare_data(spadlData)

# visualise grid:
prep_dat$bin_centers \%>\%
ggplot(aes( x  = x_center, y = y_center))+
 geom_tile(fill = "white", colour = "gray") +
 geom_text(aes( label = paste0("x:", bin_start_x, ", y:", bin_start_y))) +
 draw_pitch(c(105, 68), fill = NA)

}
