% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadlData.R
\docType{data}
\name{spadlData}
\alias{spadlData}
\title{Example Event Data set in SPADL format}
\format{
A tibble with 142,465  rows and 12 columns.
The original event data is the open La Liga Season 19/20 and 20/21 data from
StatsBomb \url{https://github.com/statsbomb/open-data}, converted to SPADL format
(unified format for event data sets from various providers, see
\url{https://arxiv.org/pdf/1802.07127.pdf}) by Grant Rhines in
his repository for an implementation of expected threat.
}
\source{
\url{https://github.com/gkrhines/xThreatR/tree/main/Statsbomb\%20Data}
}
\usage{
data(spadlData)
}
\description{
Event Data from La Liga season 2019/2020 and season 2020/2021,
publicly available from StatsBomb
}
\keyword{datasets}
