% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_dynamics.R
\name{match_dynamics}
\alias{match_dynamics}
\title{Plot match dynamics (variable over time)}
\usage{
match_dynamics(
  df,
  binsize,
  type,
  outcome_names = list(pass = NA, shot = c("Saved", "Goal", "Blocked", "Off T")),
  ylim = c(0, 100)
)
}
\arguments{
\item{df}{The event data set}

\item{binsize}{The bin size (in minutes) within which to compute the variable of interest}

\item{type}{Name of the variable for which to compute the dynamics. Must be one of
\code{Possession}, \code{Pass}, \code{Pressure}, \code{xg}.}

\item{outcome_names}{A list containing outcome names for different variables which are then
considered in the statistics, e.g. \code{outcome_names = list(pass = NA, shot = "Saved", "Goal")} only plots
vertical lines for shots that were either saved or a goal, and only considers passes that were complete
(which is NA in event data)}

\item{ylim}{Adjusts the limits of y-axis}
}
\value{
a ggplot object
}
\description{
Plot match dynamics (variable over time)
}
\examples{

FreeComp <- StatsBombR::FreeCompetitions()
FreeMatch <-  StatsBombR::FreeMatches(FreeComp \%>\% filter(season_name == "2012/2013"))
# filter match with id 18240
FreeMatch <- FreeMatch \%>\% filter(match_id == 18240)
eventData <- StatsBombR::get.matchFree(FreeMatch)
eventData <- StatsBombR::allclean(eventData)

match_dynamics(eventData, binsize = 5, type  = "Possession", outcome_names = NA)
match_dynamics(eventData, binsize = 5, type  = "Pass", outcome_names = NA)
match_dynamics(eventData, binsize = 5, type  = "Pressure", outcome_names = NA)
match_dynamics(eventData, binsize = 5, type  = "xg", outcome_names = NA)

}
